Trigger handleLoadPackage() earlier
Currently, handleLoadPackage() is triggered from LoadedApk.makeApplication().
However, this is later in the application process lifecycle than it should
be. A module which wants to hook something in the early stages of the
process will not be able to do so. Also the resources for the package used
to be loaded before the hook was triggered.

This commit moves the trigger to ActivityThread.handleBindApplication(). I
have tested that this function is called at the beginning of every
application process, also for secondary processes, services, receivers
etc. This happens shortly after the fork, but late enough to get almost
all necessary information about the new process. The only missing part is
the classloader, however it could be retrieved after calling
getPackageInfoNoCheck(). I checked how this harms the initialization and
think it's ok to call it earlier.

The special case for the system_server is triggered from ServerThread.run
as handleBindApplication() isn't called for it. The former trigger
happened way later. This also means that no system services are available
with the new location. As most apps will mainly just set up their hooks at
this time, that should be fine. We could add additional hook points later
(e.g. for when the system is ready, but before starting the applications).
On the other hand, any module can do this for itself.

This also adds ApplicationInfo, Configuration and CompatibilityInfo fields
for the callback parameter. These are no deep copies, but references.
Might be useful to be able to change something there before the
application is initialized.