add process name to LoadPackageParam, remove config and compat again
I tried to change the config at that time - it resulted in constant
reloads. Better do it in updateConfiguration() as AppSettings already
does.

Instead, pass the process name to the callback. The same package might
have two or more processes and the callback will be called for all of
them. The process name might help to target the correct one.

It's also possible that multiple packages are in the same process. This
case is now also handled. The same callback is used. With the classloader,
modules should be able to hook functions as usual. If they need to do
something only once per thread, they can check isFirstApplication. There
is an even more edge case when an app explicitely creates the context for
another packages and could potentially execute the code in its own context
(with its own uid etc). It would be no big deal to add support for this as
well, but I think this edge is very small in wildlife.