Provide an implementation of SharedPreferences
This assures that we always have the same behaviour. Especially, we can
be sure that there won't be write access.

The methods in AndroidAppHelper are deprecated with this as modules can
simply create a new instance of our class and call the reload() method
when required. This is better OO than calling a static method with the
object and there is no reason to prevent direct access to the method.