don't add XposedBridge.jar to classpath for every app
It's not required for modules to work. The original idea was to allow
Activitys coming with the module to use e.g. the logging functions.
However, this won't work when Xposed is not active, which could result in
crashes.

For such obvious usages, it might have been acceptable to add some checks
and try..catch blocks, however the same happens for indirect usages.
For example: Activity accesses a variable or method from the class which
implements an Xposed interface like IXposedHookLoadPackage. As long as
Xposed is active - and very likely, developers will have it enabled - this
worked fine. But it will crash when Xposed is not active.

The removal of this feature makes it easier for developers to spot such
dependencies.