Bump version to 30
Why moving from 2.1.4 to 30?

Version numbers can be anything. Using the x.y.z format is a tradional way
of indicating the amount and importance of changes between two versions to
end-users. However, Xposed is more than XposedBridge, there is also the
native binary and the installer. The state of this set of components is
better shown by the version number of the installer, which is displayed in
the "About" screen. That currently makes two very similar x.y.z versions
which can easily be confusing.

Apart from the one mentioned above, there is no big advantage of writing
the version as x.y.z. On the contrary, simple integers are much easier to
compare - and that's what this version number is mainly used for. Checking
whether the latest JAR is installed and checking "xposedminversion" of
modules. And the best thing: Both formats are compatible with the current
version comparison implementation. 30 is like 30.0.0, which is clearly
bigger than 2.1.4. And 2.1.4 can be seen as 2, which is lower than 30.

The lowest possible number would have been 3, but as this seems too small
and wouldn't reflect the many iterations XposedBridge already went
through. Therefore, it's 30 now.