v34: Workaround for Paranoid-based ROMs
PA added some additional instructions to the Resources constructor which
they need for their "hybrid" feature. However, this is not present in the
constructor for the Resources.getSystem() instance. As Xposed creates an
XResources instance and uses the "normal" super constructor for this,
hybrid settings will affect more parts of the ROM than usual (e.g. the
lockscreen).

Unfortunately, the default (system) constructor is private, so Xposed
can't use it. Therefore one of the PA-specific commands is nuked during
the creation of the system resources and is restored afterwards.