v35: Hack to get Xposed fully working on MIUI
MIUI uses a MiuiResources class, which is a subclass of Resources. This is
very similar to how Xposed replaces the value returned by the method
getTopLevelResources() with an XResources instance. As MIUI tries to cast
this value to MiuiResources later, we used to end up in a bootloop.

The new approach is to make XResources extend MiuiResources to make the
cast work. This also requires a hack in the binary to remove the "final"
flag of MiuiResources.

In order to retain compatibility with other ROMs (without having to
maintain two branches), Xposed includes a very simple MiuiResources class
itself. Except for the constructor, this shouldn't result in any
additional instructions for those ROMs. On Miui ROMs, this class is never
seen because it is hidden by the implementation that comes with the ROM.