Compatibility to avoid bootloops
Added back the previous signature of hookMethodNative() to avoid crashes
with app_process < v43. One devices with S-On or similar, it could happen
that an older app_process become active again.

Newer versions no longer fail if a native method could not be resolved,
therefore this compatibility method will be removed soon.