Trigger layout hooks also if the layout was included
Previously, the callback was only triggered for the root layout because
only for this the inflate() function is called.

The solution is actually a bit tricky because the XML parser for the
included layout is a local variable. But as we hook the creation of the
parser, we can check if it was created in inside parseInclude() and
remember the details for it. This is done with a LinkedList (for recursive
calls) inside a ThreadLocal.

It's assumed that the included view is the last one. This is almost always
the case, except if it's a <merge> layout, which doesn't have its own root
view but merges into the parent. In this case, it's as easy as calling
liparam.view.getParent() to get the container into which all the views
have been merged.