v44: Simple switch to turn off resource hooking
Useful if a ROM has a theming engine which is not (yet) supported by
Xposed. Will break some modules, but at least some others will work. The
most obvious breakage (module implements IXposedHookInitPackageResources,
which isn't called) is prevented by checking for this situation and not
loading the class at all.