Hook all getTopLevelResources() methods
There are many different ROM versions and theming engines which have
different versions of getTopLevelResources(). I'm tired of explicitely
coding all the different possibilities, so let's hook all of them.

Instead of constructing a ResourcesKey from the parameters manually, we
just take the one which is created by Android during the call to
getTopLevelResources(). At the moment, this is the only place where such
objects are created, but even if that changes, the implementation should
be able to deal with it.

In SDK18 and below, there were two different getTopLevelResources()
methods, one calling the other. The inner one will win and return the
replaced XResources object to the outer one.