Simplify calls to handleInitPackageResources()
The old code had a little bug anyway (null vs. "") and I have never seen
that the package name wasn't available. If such a situation occurs, that
should be fixed instead of delaying the initialization. Also this can be
called for resDir == null, otherwise the ResourcesKey would throw an NPE.