Dirty clone Resources to XResources
The native method is able to clone any object (even if it doesn't
implement Cloneable). The new object can be allocated with enough memory
for a subclass of the original class. All the original fields are cloned,
and any additional fields are zeroed out (so they should be set
explicitely). This makes use of the fact that inherited fields have the
same memory layout in the subclass and additional fields are stored on top
of (= after) the inherited ones.

We use this to create the XResources object without having to worry about
ROM-specific fields or constructors.

Requires binary version 54.