Add a cache for resource replacement status
Store whether there are replacements for a certain group of IDs to speed
up getReplacement() significantly. The group is calculated from the most
significant bits of the ID, which should usually avoid false positives.
If the cache says that there are no replacements, then there are
definitely none and the more expensive lookup in the SparseArray can be
skipped. Otherwise, there *might* be replacements. The lookup still
happens as before and might return null. Therefore, false positives in the
cache won't cause bad behavior, just longer runtimes (but only very little
more than the runtime without this commit for every single resource).