Fix delayed configuration update for forwarded resources
addActiveResource() created the ResourcesKey with the current
Configuration in Android 4.2 and higher.
However, this configuration is actually used to override the system
configuration, it should be "null" unless explicitely requested otherwise.

Due to this bug, resource replacements that forward to the module's
resources were loaded with an outdated configuration. For some reason, the
new configuration was applied on the next configuration update, so it was
always one step behind.

Now the system configuration is applied correctly.