Hard-limit log file size to 5 MB
In all cases I have seen so far, huge log files were caused by
incompatible modules, writing thousands of error messages.
Still, people blame it on Xposed and ask for an option to disable logs
(which obviously won't fix anything). This is a compromise and users
should probably be warned about this in the UI.

Fixes #17