No longer write the log to file from Java code
Due to SELinux, it's pretty much impossible to open file for writing in
the Zygote process and keep writing to it from application and system
processes. On 64-bit ROMs, it's even more complicated because there are
two Zygote processes.

Messages will still be written to logcat and are captured from there in
native coding.