Differentiate between the two Zygote processes on 64-bit ROMs
Only one of these processes starts the system_server process, so it only
makes sense to place some hooks or trigger certain actions for this
primary process.

Modules can check this in initZygote() via a new flag in the parameters.
Generally, it's strongly recommended to place any hooks for system_server
in handleLoadPackage() for the "android" package instead.