Work around SELinux restrictions
This commit introduces various classes and helpers to access files even
when SELinux is active and enforcing. Depending on the process/context
(Zygote, system_server, normal app), different approaches have to be used,
most of them realized via native services.

The intention is to abstract the very complex logic and the choice of the
right service by providing a common interface for them. XSharedPreferences
is adapted to handle this in the background, so modules loading their
preferences via this class shouldn't need to be updated at all (provided
they store the preferences world-readable, which has been a requirement
already in most cases).

Modules are free to use the SELinuxHelper class and the services retrieved
from it to access files in /data/data/ themselves. The API is pretty much
finished, although breaking changes are still possible until the first
stable version of Xposed for Lollipop has been published.