Hook getTopLevelThemedResources()
CM11 has introduced this method with Change-Id
If04e82a13250f520d70b58991586c8ce38f0ecb1.

Especially on CM12+, the SystemUI (and possibly other parts of the system)
use this method to retrieve a Resources object. We need to hook it and
replace the result with an XResources object.