Clean up unhooking methods
Modules should always use the IXUnhook instances that are returned from
the hooking methods. Also hooking LoadPackage and InitPackageResources
should be done using the interfaces for the method class, not with
stand-alone classes. This will need further cleanup in the future.

Remove/hide unused methods, deprecate unhookMethod() to give the very
few modules using it a chance to switch.