Don't force apps to compile during the boot process
This change reverts the following commits:
  01c324454bb0e7eba82ae56895b04fb571ba5073
  f5967d50455e4103f6353446b376b3806ba5e429

They have become obsolete by a change in Xposed's ART libraries which
allows the system to use the dex from .odex files in interpret-only mode
(i.e. the compiled code is ignored).

This works out-of-the-box on stock ROMs where the image checksums match
(because the image hasn't been recompiled) and no Xposed version check
prevents the usage of the .odex file.