Warn developers when Instant Run is enabled in Android Studio
Instant Run replaces the application with a bootstrap application that
can communicate with Android Studio and dynamically exchange methods
In order to achieve this, actual classes are not stored directly in the
APK's classes.dex, but within a file called instant-run.zip in the APK.
Therefore, these classes can't be found by Xposed unless it would
simulate the same bootstrapping that takes place when the app is started.

Let's avoid the misleading ClassNotFoundException and instead print a
proper error message asking the user to disable Instant Run.