Generic handling for ROMs with custom Resources subclasses
A couple of ROMs (including Huawei's stock ROMs) use a subclass of the
Resources class for theming, just like Xposed. In the past, Xposed
couldn't inject its XResources class on such ROMs as it would cause
crashes when the ROM tries to call method that don't exist in XResources.

The trick here is to make XResources dynamically extend the ROM's
resources class. This is achieved with the following inheritance chain:

  1.  XResources
  2.   extends XResourcesSuperClass
  3.   extends <ROM resources class>
  4.   extends Resources

1 <=> 2 is defined in XposedBridge.jar
3 <=> 4 is defined in the ROM

The missing connection 2 <=> 3 is defined in a dynamically created dex
file which just contains the definition
  "XResourcesSuperClass extends <ROM resources class>"

This dex file is then wrapped into a ClassLoader and injected as parent
of XposedBridge's ClassLoader.

All of this has to happen before XResources is accessed the first time
and there mustn't be references to XResource in any classes used up to
that point, therefore XposedBridge itself should include only what's
absolutely necessary.